refactor: 💡 Rename of components
